before switch
