+++ webui/react/src/components/ResponsiveTable.module.scss
@@ -6,3 +6,123 @@
 .rightClickableCellWrapper {
   display: inline;
 }
+
+
+/* .react-resizable {
+  background-clip: padding-box;
+  position: relative;
+} */
+
+/* tr.drop-over-leftward td {
+  border-left: 2px dashed #1890ff;
+}
+tr.drop-over-rightward td {
+  border-right: 2px dashed #1890ff;
+} */
+
+.drop-over-leftward {
+  border-left: 2px dashed #1890ff !important;
+}
+.drop-over-rightward {
+  border-right: 2px dashed #1890ff !important;
+}
+.ant-table-cell {
+  cursor: 'move';
+  user-select: none;
+}
+.ant-table-filter-trigger {
+  cursor: pointer
+}
+.ant-table-column-sorter {
+  cursor: pointer;
+  z-index: 1;
+}
+.ant-notification-notice {
+  display: none;
+}
+.react-resizable-handle {
+  border-radius: 4px;
+  bottom: 0;
+  cursor: col-resize;
+  height: 100%;
+
+  /* margin-left: 4px; */
+  position: absolute;
+  right: -2.5px;
+  user-select: none;
+  width: 4px;
+  z-index: 1;
+}
+.react-resizable-handle:hover {
+  background: rgba(68, 142, 247, 0.4);
+}
+
+/* .dragHandler:hover {
+  background-color: #ccc;
+  cursor: move;
+}
+.react-resizable {
+  position: relative;
+} */
+
+/* .react-resizable-handle {
+  background-origin: content-box;
+  background-position: bottom right;
+  background-repeat: no-repeat;
+  box-sizing: border-box;
+  height: 20px;
+  padding: 0 3px 3px 0;
+  position: absolute;
+  width: 20px;
+}
+.react-resizable-handle-sw {
+  bottom: 0;
+  cursor: sw-resize;
+  left: 0;
+  transform: rotate(90deg);
+}
+.react-resizable-handle-se {
+  bottom: 0;
+  cursor: se-resize;
+  right: 0;
+}
+.react-resizable-handle-nw {
+  cursor: nw-resize;
+  left: 0;
+  top: 0;
+  transform: rotate(180deg);
+}
+.react-resizable-handle-ne {
+  cursor: ne-resize;
+  right: 0;
+  top: 0;
+  transform: rotate(270deg);
+}
+.react-resizable-handle-w,
+.react-resizable-handle-e {
+  cursor: ew-resize;
+  margin-top: -10px;
+  top: 50%;
+}
+.react-resizable-handle-w {
+  left: 0;
+  transform: rotate(135deg);
+}
+.react-resizable-handle-e {
+  right: 0;
+  transform: rotate(315deg);
+}
+.react-resizable-handle-n,
+.react-resizable-handle-s {
+  cursor: ns-resize;
+  left: 50%;
+  margin-left: -10px;
+}
+.react-resizable-handle-n {
+  top: 0;
+  transform: rotate(225deg);
+}
+.react-resizable-handle-s {
+  bottom: 0;
+  transform: rotate(45deg);
+} */
